chore: Move Cards inside CardsWrapper
