+++ src/components/EndedCardsWrapper/EndedCard/Votes/index.tsx
