+++ src/components/EndedCardsWrapper/EndedCard/Votes/v.module.css
