+++ src/components/EndedCardsWrapper/EndedCard/ec.module.css
