+++ src/components/EndedCardsWrapper/EndedCard/index.tsx
